Changes in e2e tests: - Move all methods that create dummy data into TestBase - Unify given_an_ like methods into given_a_page() in TestBase - Introduce when_logs_into_admin_page() test method
