+++ blog/tests/e2e/test_main_page.py
@@ -7,7 +7,7 @@ class TestMainPage(TestBase):
         """Tests that main page displays list of existing articles.
         """
         self.create_dummy_articles()
-        self.given_a_main_page()
+        self.given_a_page("Main")
         self.when_vist_this_page("My Personal Blog")
         self.then_i_see_list_of_articles()
     
@@ -24,14 +24,3 @@ class TestMainPage(TestBase):
         """
         articles = self.browser.find_elements(By.CSS_SELECTOR, ".article-link")
         self.assertTrue(len(articles) > 0)
-    
-    def create_dummy_articles(self):
-        """Creates dummy articles.
-        """
-        test_articles = [
-            {"title": "Test Article 1", "content": "Test Article 1"},
-            {"title": "Test Article 2", "content": "Test Article 2"},
-            {"title": "Test Article 3", "content": "Test Article 3"}
-        ]
-        for article in test_articles:
-            self.create_dummy_article(article)
